Update release and package scripts (#93)
Updates the release and package scripts to resolve errors during `yarn publish:all` flow. With the previous implementation, `npm` would attempt to publish each package twice, seemingly because we defined a package script named `publish`. This script has been renamed to `publish:package` so as to not collide with the `npm publish` command.

As part of this change, the `prepublishOnly` script has been removed for each package as it performs redundant work when packages are published via the `publish:all` command. This is potentially ~unsafe in cases were single packages are published directly, but I don't foresee us doing so until we migrate to a different publishing flow entirely (ref: #66).

Finally, the deprecated `setup` package script is also removed.